code/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
atlassian-python-api==3.41.9
attrs==23.2.0
azure-ai-contentsafety==1.0.0b1
azure-ai-formrecognizer==3.3.2
azure-common==1.1.28
azure-core==1.29.7
azure-functions==1.18.0
azure-identity==1.15.0
azure-search-documents==11.4.0b8
azure-storage-blob==12.19.0
azure-storage-queue==12.9.0
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cryptography==41.0.7
cssselect2==0.7.0
cycler==0.12.1
dataclasses-json==0.6.3
Deprecated==1.2.14
distro==1.9.0
docx2txt==0.8
fake-useragent==1.4.0
filelock==3.13.1
Flask==3.0.1
fonttools==4.47.2
freetype-py==2.4.0
frozenlist==1.4.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.41
google-api-core==2.15.0
google-auth==2.26.2
googleapis-common-protos==1.62.0
grandalf==0.8
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.2
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.1.2
langchain-community==0.0.14
langchain-core==0.1.14
langchain-openai==0.0.3
langchainhub==0.1.14
langsmith==0.0.83
lxml==5.1.0
markdown-it-py==3.0.0
markdownify==0.11.6
MarkupSafe==2.1.4
marshmallow==3.20.2
matplotlib==3.8.2
mdurl==0.1.2
msal==1.26.0
msal-extensions==1.1.0
msrest==0.7.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.3
oauthlib==3.2.2
openai==1.9.0
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.13
packaging==23.2
pandas==1.5.1
pdf2image==1.17.0
pgvector==0.2.4
pillow==10.2.0
plotly==5.18.0
pluggy==1.3.0
portalocker==2.8.2
protobuf==4.25.2
psutil==5.9.8
psycopg2==2.9.9
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.25.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
pypdf==4.0.0
pytesseract==0.3.10
pytest==7.4.4
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
regex==2023.12.25
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rlPyCairo==0.3.0
rpds-py==0.17.1
rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
streamlit==1.30.0
svglib==1.5.1
tenacity==8.2.3
tesseract==0.1.3
threadpoolctl==3.2.0
tiktoken==0.5.2
tinycss2==1.2.1
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
transformers==4.36.2
types-requests==2.31.0.20240125
typing-inspect==0.9.0
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
webencodings==0.5.1
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
